Break the unit test into files. Each file for one test class
